dashboard/requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
paho-mqtt>=1.6.1
numpy>=1.24.0
